+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 **TR4**
 - added reflections
+- added water droplets dripping off Lara after she leaves water (also enabled in TR3)
 - fixed the camera snapping to elevation and angle changes in instances such as opening floor trapdoors
 
 **Lua**